Renamed component files to PascalCase
