+++ src/components/CoverImage.tsx
