+++ src/components/PostHeader.tsx
@@ -1,5 +1,5 @@
 import Avatar from './avatar';
-import CoverImage from 'src/components/cover-image';
+import CoverImage from 'src/components/CoverImage';
 import type Author from 'src/interfaces/author';
 import { formatDate } from 'src/lib/date';
